refactor worker handlers into submodules
