+++ app/amochats.py
@@ -1,0 +1,8 @@
+async def send_text_from_manager(*args, **kwargs):
+    return None
+
+async def ensure_chat_created(*args, **kwargs):
+    return None
+
+async def send_text_from_client(*args, **kwargs):
+    return None
