+++ app/config.py
@@ -1,0 +1,6 @@
+class Settings:
+    def __getattr__(self, name):
+        return ""
+
+
+settings = Settings()
