+++ app/logging_setup.py
@@ -1,0 +1,2 @@
+def setup_logging(level: str) -> None:
+    pass
