+++ app/oauth2.py
@@ -1,0 +1,2 @@
+async def ensure_fresh_access(*args, **kwargs):
+    return ""
